+++ src/core/prelude/object/iterators/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.35.2 (2021-03-12)
+
+#### :boom: Breaking Change
+
+* Marked `notOwn` as deprecated. Use `propsToIterate` instead.
+* Marked `withDescriptor` as deprecated. Use `passDescriptor` instead.
+
 ## v3.35.0 (2021-03-04)
 
 #### :rocket: New Feature